Project_1/requirements.txt
fastapi==0.85.0
pytest==7.1.3
uvicorn==0.18.3
pydantic~=1.9.2
SQLAlchemy
python-dotenv
psycopg2
passlib
python-jose
